+++ packages/client/src/LensClient.ts
@@ -14,7 +14,6 @@ import {
   Momoka,
   Nfts,
   Notifications,
-  Poaps,
   Profile,
   Publication,
   Revenue,
@@ -156,13 +155,6 @@ export class LensClient {
    */
   get notifications(): Notifications {
     return new Notifications(this.context, this._authentication);
-  }
-
-  /**
-   * The Poaps module
-   */
-  get poaps(): Poaps {
-    return new Poaps(this.context, this._authentication);
   }
 
   /**