+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lens-protocol/react
 
+## 2.0.0-alpha.38
+
+### Patch Changes
+
+- 8dcaeda61: **chore:** remove unused local-only field
+- 880fb5de3: **feat:** supports Unknown Reference modules
+- Updated dependencies [8dcaeda61]
+- Updated dependencies [880fb5de3]
+  - @lens-protocol/api-bindings@0.11.0-alpha.34
+  - @lens-protocol/domain@0.11.0-alpha.27
+  - @lens-protocol/blockchain-bindings@0.10.0-alpha.28
+
 ## 2.0.0-alpha.37
 
 ### Minor Changes